pyproject.toml
[tool.poetry]
name = "understanding_flow_robustness"
version = "0.1.0"
description = "Official repository for Towards Understanding Adversarial Robustness of Optical Flow Networks (CVPR 2022)"
authors = ["Simon Schrodi <schrodi@cs.uni-freiburg.de>"]
license = "GNU"

[tool.poetry.dependencies]
python = "3.8"
gdown = "^4.4.0"
einops = "^0.4.1"
joblib = "^1.1.0"
tqdm = "^4.63.0"
path = "^16.4.0"
Pillow = "^9.0.1"
blessings = "^1.7"
scipy = "^1.8.0"
tensorboard = "^2.8.0"
progressbar2 = "^4.0.0"
pypng = "^0.0.21"
scikit-image = "^0.19.2"
opencv-python = "^4.5.5"
matplotlib = "^3.5.1"
imagecorruptions = "^1.1.2"
tensorboardX = "^2.5"
toml = "^0.10.2"
termcolor = "^1.1.0"
umap-learn = "^0.5.2"
pandas = "^1.4.1"
pre-commit = "^2.17.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0"
pre-commit = "^2.17"
pytest = "^6.2.5"
mypy = "^0.930"
mkdocs-material = "^8.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ['py38']

[tool.isort]
profile = 'black'
line_length = 90

[tool.pytest.ini_options]
addopts = "--basetemp ./tests_tmpdir"

[tool.mypy]
files = ["neps"]
namespace_packages = true
follow_imports = "skip"
ignore_missing_imports = true
show_error_codes = true
incremental = false
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
  "neps.optimizers.bayesian_optimization.models.*",
  "neps.optimizers.bayesian_optimization.kernels.*",
  "neps.optimizers.bayesian_optimization.acquisition_functions.*",
  "neps.optimizers.bayesian_optimization.acquisition_function_optimization.*",
  "neps.search_spaces.graph_dense.*",
]
ignore_errors = true

[tool.pylint.messages_control]
disable = 'all'
enable = [
    'invalid-characters-in-docstring','wrong-spelling-in-comment','wrong-spelling-in-docstring','not-in-loop',
    'function-redefined','continue-in-finally','abstract-class-instantiated','star-needs-assignment-target',
    'duplicate-argument-name','return-in-init','too-many-star-expressions','nonlocal-and-global',
    'return-outside-function','return-arg-in-generator','invalid-star-assignment-target','bad-reversed-sequence',
    'nonexistent-operator','yield-outside-function','init-is-generator','nonlocal-without-binding','lost-exception',
    'assert-on-tuple','dangerous-default-value','duplicate-key','useless-else-on-loop','expression-not-assigned',
    'confusing-with-statement','unnecessary-lambda','pointless-statement','unnecessary-pass','unreachable','eval-used',
    'exec-used','using-constant-test','deprecated-lambda','blacklisted-name','misplaced-comparison-constant',
    'singleton-comparison','unneeded-not','ßconsider-iterating-dictionary','consider-using-enumerate','empty-docstring',
    'unidiomatic-typecheck','condition-evals-to-constant','consider-using-generator','nan-comparison',
    'consider-using-min-max-builtin','consider-using-with','invalid-all-format','consider-using-dict-items',
    'deprecated-decorator','forgotten-debug-statement','unspecified-encoding','useless-with-lock',
    'use-implicit-booleaness-not-comparison','not-async-context-manager','yield-inside-async-function',
    'await-outside-async','invalid-unary-operand-type','unsupported-binary-operation','not-callable',
    'redundant-keyword-arg','assignment-from-no-return','assignment-from-none','not-context-manager',
    'repeated-keyword','missing-kwoa','no-value-for-parameter','invalid-sequence-index','invalid-slice-index',
    'unexpected-keyword-arg','unsupported-membership-test','unsubscriptable-object','bad-except-order',
    'catching-non-exception','bad-exception-context','notimplemented-raised','raising-bad-type','raising-non-exception',
    'misplaced-bare-raise','duplicate-except','nonstandard-exception','binary-op-exception','bare-except',
    'raise-missing-from','consider-using-namedtuple-or-dataclass','consider-using-tuple','bad-open-mode',
    'redundant-unittest-assert','boolean-datetime','deprecated-methodimport-error','import-self','reimported',
    'relative-import','deprecated-module','wildcard-import','misplaced-future','cyclic-import','wrong-import-position',
    'ungrouped-imports','multiple-imports','simplifiable-condition','len-as-condition',
    'unpacking-non-sequence','invalid-all-object','unbalanced-tuple-unpacking','undefined-variable',
    'undefined-all-variable','used-before-assignment','cell-var-from-loop','global-variable-undefined',
    'redefined-builtin','redefine-in-handler','unused-import','unused-argument','unused-wildcard-import',
    'unused-variable','global-variable-not-assigned','undefined-loop-variable','global-statement',
    'global-at-module-level','format-needs-mapping','truncated-format-string','missing-format-string-key',
    'mixed-format-string','too-few-format-args','bad-str-strip-call','too-many-format-args','bad-format-character',
    'format-combined-specification','bad-format-string-key','bad-format-string','missing-format-attribute',
    'missing-format-argument-key','unused-format-string-argument','unused-format-string-key','invalid-format-index',
    'f-string-without-interpolation','use-maxsplit-arg','anomalous-unicode-escape-in-string',
    'anomalous-backslash-in-string','redundant-u-string-prefix','format-string-without-interpolation',
    'simplifiable-if-statement','logging-format-truncated','logging-too-few-args','logging-too-many-args',
    'logging-unsupported-format','not-an-iterable','not-a-mapping','use-sequence-for-iteration','bad-indentation',
    'unnecessary-semicolon','missing-final-newline','mixed-line-endings','multiple-statements','trailing-newlines',
    'trailing-whitespace','unexpected-line-ending-format','superfluous-parens','access-member-before-definition',
    'method-hidden','assigning-non-slot','duplicate-bases','inconsistent-mro','inherit-non-class','invalid-slots',
    'invalid-slots-object','no-method-argument','no-self-argument','unexpected-special-method-signature',
    'non-iterator-returned','invalid-length-returned','protected-access','attribute-defined-outside-init',
    'abstract-method','bad-staticmethod-argument','non-parent-init-called','super-init-not-called',
    'no-classmethod-decorator','no-staticmethod-decorator','no-self-use','bad-classmethod-argument',
    'bad-mcs-classmethod-argument','bad-mcs-method-argument','method-check-failed','invalid-bool-returned',
    'invalid-index-returned','invalid-repr-returned','invalid-str-returned','invalid-bytes-returned',
    'invalid-hash-returned','invalid-length-hint-returned','invalid-format-returned','invalid-getnewargs-returned',
    'invalid-getnewargs-ex-returned','super-with-arguments','deprecated-class','invalid-class-object',
    'unused-private-member',
]
